Modifications in projects organisation : - Now target .NET Standard 2.0 and NET Framework 4.0 and 4.5 - Put Unit Tests in a specific project - NuGet package is generated on build
