+++ CSharpFITS_v1.1.sln
@@ -1,0 +1,36 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.27703.2026
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CSharpFITS_v1.1", "CSharpFITS_v1.1\CSharpFITS_v1.1.csproj", "{93C93A99-D18F-495E-8AC9-6B6F6F7F273E}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{0C07CCD3-FF69-4555-A6AF-73603B6A2D1A}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CSharpFITS.Test", "tests\CSharpFITS.Test\CSharpFITS.Test.csproj", "{B6B22F07-02B3-40E3-8498-83BA3AC49712}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{93C93A99-D18F-495E-8AC9-6B6F6F7F273E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{93C93A99-D18F-495E-8AC9-6B6F6F7F273E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{93C93A99-D18F-495E-8AC9-6B6F6F7F273E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{93C93A99-D18F-495E-8AC9-6B6F6F7F273E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B6B22F07-02B3-40E3-8498-83BA3AC49712}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B6B22F07-02B3-40E3-8498-83BA3AC49712}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B6B22F07-02B3-40E3-8498-83BA3AC49712}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B6B22F07-02B3-40E3-8498-83BA3AC49712}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{B6B22F07-02B3-40E3-8498-83BA3AC49712} = {0C07CCD3-FF69-4555-A6AF-73603B6A2D1A}
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {EDD6DA51-A08C-4CBC-9C53-1C6C661172B4}
+	EndGlobalSection
+EndGlobal
