+++ CSharpFITS_v1.1/CSharpFITS_v1.1.csproj
@@ -1,126 +1,22 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003" ToolsVersion="14.0">
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
   <PropertyGroup>
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{93C93A99-D18F-495E-8AC9-6B6F6F7F273E}</ProjectGuid>
-    <OutputType>Library</OutputType>
-    <NoStandardLibraries>false</NoStandardLibraries>
-    <AssemblyName>CSharpFITS_v1.1</AssemblyName>
-    <RootNamespace>CSharpFITS_v1._1</RootNamespace>
-    <StartupObject>
-    </StartupObject>
-    <TargetFrameworkVersion>v2.0</TargetFrameworkVersion>
-    <FileUpgradeFlags>
-    </FileUpgradeFlags>
-    <UpgradeBackupLocation>
-    </UpgradeBackupLocation>
-    <OldToolsVersion>2.0</OldToolsVersion>
+    <TargetFrameworks>netstandard2.0;net40;net45</TargetFrameworks>
+    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+    <PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
+    <Product>CSharpFITS</Product>
+    <Company>Virtual Observatory India</Company>
+    <Copyright>Thomas McGlynn 1997-2007, Virtual Observatory - India 2007.</Copyright>
+    <PackageLicenseUrl>https://github.com/elendil-software/csharpfits/blob/master/license.txt</PackageLicenseUrl>
+    <AssemblyVersion>1.1.1.0</AssemblyVersion>
+    <FileVersion>1.1.1.0</FileVersion>
+    <PackageTags>Astronomy,FITS,Image</PackageTags>
+    <RepositoryUrl>https://github.com/elendil-software/csharpfits</RepositoryUrl>
+    <PackageProjectUrl>https://github.com/elendil-software/csharpfits</PackageProjectUrl>
+    <Description></Description>
+    <Version>1.1.1</Version>
+    <Authors>Virtual Observatory - India</Authors>
+    <PackageId>CSharpFITS</PackageId>
   </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
-    <DebugSymbols>true</DebugSymbols>
-    <DebugType>full</DebugType>
-    <Optimize>false</Optimize>
-    <OutputPath>.\bin\Debug\</OutputPath>
-    <DefineConstants>DEBUG;TRACE</DefineConstants>
-    <DocumentationFile>bin\Debug\CSharpFITS_v1.1.XML</DocumentationFile>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-    <NoWarn>1591, 1573, 1572, 1570, 1584, 1658, 1587</NoWarn>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-    <DebugType>pdbonly</DebugType>
-    <Optimize>true</Optimize>
-    <OutputPath>.\bin\Release\</OutputPath>
-    <DefineConstants>TRACE</DefineConstants>
-    <DocumentationFile>bin\Release\CSharpFITS_v1.1.XML</DocumentationFile>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-  </PropertyGroup>
-  <ItemGroup>
-    <Reference Include="nunit.framework, Version=2.6.4.14350, Culture=neutral, PublicKeyToken=96d09a1eb7f44a77, processorArchitecture=MSIL">
-      <HintPath>packages\NUnit.2.6.4\lib\nunit.framework.dll</HintPath>
-      <Private>True</Private>
-    </Reference>
-    <Reference Include="System" />
-    <Reference Include="System.Data" />
-    <Reference Include="System.Drawing" />
-    <Reference Include="System.Xml" />
-  </ItemGroup>
-  <ItemGroup>
-    <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="src\nom\tam\fits\AsciiTable.cs" />
-    <Compile Include="src\nom\tam\fits\AsciiTableHDU.cs" />
-    <Compile Include="src\nom\tam\fits\BadHeaderException.cs" />
-    <Compile Include="src\nom\tam\fits\BasicHDU.cs" />
-    <Compile Include="src\nom\tam\fits\BinaryTable.cs" />
-    <Compile Include="src\nom\tam\fits\BinaryTableHDU.cs" />
-    <Compile Include="src\nom\tam\fits\Data.cs" />
-    <Compile Include="src\nom\tam\fits\Fits.cs" />
-    <Compile Include="src\nom\tam\fits\FitsDate.cs" />
-    <Compile Include="src\nom\tam\fits\FitsElement.cs" />
-    <Compile Include="src\nom\tam\fits\FitsException.cs" />
-    <Compile Include="src\nom\tam\fits\FitsFactory.cs" />
-    <Compile Include="src\nom\tam\fits\FitsHeap.cs" />
-    <Compile Include="src\nom\tam\fits\FitsUtil.cs" />
-    <Compile Include="src\nom\tam\fits\Header.cs" />
-    <Compile Include="src\nom\tam\fits\HeaderCard.cs" />
-    <Compile Include="src\nom\tam\fits\HeaderCardException.cs" />
-    <Compile Include="src\nom\tam\fits\HeaderOrder.cs" />
-    <Compile Include="src\nom\tam\fits\ImageData.cs" />
-    <Compile Include="src\nom\tam\fits\ImageHDU.cs" />
-    <Compile Include="src\nom\tam\fits\RandomGroupsData.cs" />
-    <Compile Include="src\nom\tam\fits\RandomGroupsHDU.cs" />
-    <Compile Include="src\nom\tam\fits\StreamedBinaryTableHDU.cs" />
-    <Compile Include="src\nom\tam\fits\TableData.cs" />
-    <Compile Include="src\nom\tam\fits\TableHDU.cs" />
-    <Compile Include="src\nom\tam\fits\TruncatedFileException.cs" />
-    <Compile Include="src\nom\tam\fits\UndefinedData.cs" />
-    <Compile Include="src\nom\tam\fits\UndefinedHDU.cs" />
-    <Compile Include="src\nom\tam\image\ImageTiler.cs" />
-    <Compile Include="src\nom\tam\util\ActualBufferedStream.cs" />
-    <Compile Include="src\nom\tam\util\AdapterStream.cs" />
-    <Compile Include="src\nom\tam\util\ArrayDataIO.cs" />
-    <Compile Include="src\nom\tam\util\ArrayFuncs.cs" />
-    <Compile Include="src\nom\tam\util\BufferedDataStream.cs" />
-    <Compile Include="src\nom\tam\util\BufferedFile.cs" />
-    <Compile Include="src\nom\tam\util\ByteFormatter.cs" />
-    <Compile Include="src\nom\tam\util\ByteParser.cs" />
-    <Compile Include="src\nom\tam\util\ByteRenderer.cs" />
-    <Compile Include="src\nom\tam\util\ColumnTable.cs" />
-    <Compile Include="src\nom\tam\util\ConfigStream.cs" />
-    <Compile Include="src\nom\tam\util\Cursor.cs" />
-    <Compile Include="src\nom\tam\util\DataTable.cs" />
-    <Compile Include="src\nom\tam\util\DefaultValueHashtable.cs" />
-    <Compile Include="src\nom\tam\util\HashedList.cs" />
-    <Compile Include="src\nom\tam\util\HeapStream.cs" />
-    <Compile Include="src\nom\tam\util\PrimitiveInfo.cs" />
-    <Compile Include="src\nom\tam\util\RandomAccess.cs" />
-    <Compile Include="src\nom\tam\util\RowSource.cs" />
-    <Compile Include="src\nom\tam\util\StringToByteStream.cs" />
-    <Compile Include="src\nom\tam\util\SupportClass.cs" />
-    <Compile Include="src\nom\tam\util\TableException.cs" />
-    <Compile Include="src\nom\tam\util\Troolean.cs" />
-    <Compile Include="src\nom\tam\util\TruncationException.cs" />
-    <Compile Include="tests\nom\tam\fits\AsciiTableTest.cs" />
-    <Compile Include="tests\nom\tam\fits\BinaryTableTest.cs" />
-    <Compile Include="tests\nom\tam\fits\FitsCopyTest.cs" />
-    <Compile Include="tests\nom\tam\fits\FitsReaderTest.cs" />
-    <Compile Include="tests\nom\tam\fits\HeaderCardTest.cs" />
-    <Compile Include="tests\nom\tam\fits\HeaderTest.cs" />
-    <Compile Include="tests\nom\tam\fits\ImageTest.cs" />
-    <Compile Include="tests\nom\tam\fits\RandomGroupsTest.cs" />
-    <Compile Include="tests\nom\tam\fits\TestDate.cs" />
-    <Compile Include="tests\nom\tam\fits\TilerTest.cs" />
-    <Compile Include="tests\nom\tam\util\ArrayFuncsTester.cs" />
-    <Compile Include="tests\nom\tam\util\BufferedFileTester.cs" />
-    <Compile Include="tests\nom\tam\util\ByteFormatParseTester.cs" />
-    <Compile Include="tests\nom\tam\util\HashedListTester.cs" />
-    <Compile Include="tests\sharpcap\TestFitsWriter.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="packages.config" />
-  </ItemGroup>
-  <Import Project="$(MSBuildBinPath)\Microsoft.CSHARP.Targets" />
-  <ProjectExtensions>
-    <VisualStudio AllowExistingFolder="true" />
-  </ProjectExtensions>
+
 </Project>
