+++ CSharpFITS_v1.1/src/nom/tam/fits/BinaryTable.cs
@@ -1,3 +1,7 @@
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("CSharpFITS.Test")]
+
 namespace nom.tam.fits
 {
     /*
