+++ tests/CSharpFITS.Test/nom/tam/util/ByteFormatParseTester.cs
